Getting hex working again
